Modified the FrameBuilder class such that we can now supply a float for the tileScale variable. For example, now we can have 32x32 px tiles from a spritesheet, and scale it to 1.5f to make them 48x48 px.

+++ src/Builders/FrameBuilder.java
@@ -19,12 +19,14 @@ public class FrameBuilder {
         if (this.delay < 0) {
             this.delay = 0;
         }
+        // when instantiated, scale is set to 1
         this.scale = 1;
         this.imageEffect = ImageEffect.NONE;
     }
 
     public FrameBuilder(BufferedImage image) {
         this.image = image;
+        // scale is still 1
         this.scale = 1;
         this.imageEffect = ImageEffect.NONE;
     }
@@ -40,8 +42,9 @@ public class FrameBuilder {
     }
 
     public FrameBuilder withScale(float scale) {
+        // convert the scale parameter to a float (which may be useless), but makes sure that its correct
         if (this.scale >= 0) {
-            this.scale = scale;
+            this.scale = (float) scale;
         } else {
             this.scale = 1;
         }
